// src/models/column-name.ts
/**
 * столбцы таблицы ПБ
 */

export enum ColumnName {
  SerialNumber = 'serialNumber',
  DocumentArrivalDate = 'documentArrivalDate',
  PassportNumber = 'passportNumber',
  Names = 'names',
  Status = 'status',
  Organization = 'organization',
  Mediator = 'mediator',
  ExpertId = 'expertId',
  StartDate = 'startDate',
  EndDate = 'endDate',
  WorkDate = 'workDate',
  Okpd2CodeId = 'okpd2CodeId',
  TnVedCodeId = 'tnVedCodeId',
}